requirements.txt
requests>=2.28.2
python-dotenv>=1.0.0
lxml>=4.9.2
Django>=4.1.7
djangorestframework>=3.14.0